# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DeciFlow API",
        "description": "API documentation for DeciFlow - Purchase Request Approval System",
        "contact": {
            "email": "support@deciflow.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/approvals/inbox": {
            "get": {
                "tags": [
                    "Approvals"
                ],
                "summary": "Get approval inbox",
                "description": "Get all requests pending approval for the current user's role",
                "operationId": "554e8660cc53cb3bac5a4b22d1450ee9",
                "responses": {
                    "200": {
                        "description": "List of pending approvals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/requests/{id}/approve": {
            "post": {
                "tags": [
                    "Approvals"
                ],
                "summary": "Approve request",
                "description": "Approve a request at current approval step",
                "operationId": "e90c3d6bb097e8947493cb7a40b01e35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Approved - looks good"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Approval failed"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/requests/{id}/reject": {
            "post": {
                "tags": [
                    "Approvals"
                ],
                "summary": "Reject request",
                "description": "Reject a request at current approval step",
                "operationId": "b6895e7b5d1ccd018582ede44147610c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Rejected - insufficient justification"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request rejected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Rejection failed"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/requests/{id}/return": {
            "post": {
                "tags": [
                    "Approvals"
                ],
                "summary": "Return request",
                "description": "Return a request to requester for revision",
                "operationId": "bd1154e8fc2650abcba19783cb729bd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Please provide more details about the vendor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Return failed"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}